# api/openapi.yaml
openapi: "3.0.3"
info:
  title: CMDB Platform API
  version: "1.0.0"
  description: Configuration Management Database Platform API
servers:
  - url: /api/v1
security:
  - BearerAuth: []
...
paths:
...
  /racks/{id}/maintenance:
    get:
      operationId: getRackMaintenance
      tags: [maintenance]
      summary: Get maintenance history for a rack
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Rack maintenance records
...
  /assets/{id}/lifecycle:
    get:
      operationId: getAssetLifecycle
      tags: [assets]
      summary: Return lifecycle status (age, warranty, depreciation, upgrade recommendation)
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Asset lifecycle
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/AssetLifecycle' }
                  meta: { $ref: '#/components/schemas/Meta' }
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    Meta:
      type: object
      properties:
        request_id:
          type: string
      required: [request_id]
...
    ErrorBody:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required: [code, message]
    ErrorResponse:
      type: object
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
        meta:
          $ref: '#/components/schemas/Meta'
      required: [error, meta]

    # ── Auth ──────────────────────────────────────────────────────────────
...
    AssetLifecycle:
      type: object
      properties:
        asset_id: { type: string, format: uuid }
        age_days: { type: integer }
        age_months: { type: integer }
        warranty_remaining_days: { type: integer, nullable: true }
        warranty_status: { type: string, enum: [active, expiring_soon, expired, unknown] }
        depreciation_pct: { type: number, format: float, nullable: true }
        upgrade_priority: { type: string, enum: [none, low, medium, high], nullable: true }
        recommended_action: { type: string, nullable: true }
      additionalProperties: true
...